pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "doc-rag-pipeline"
version = "2.0.0"
description = "Intelligent Document Q&A pipeline using local models, FAISS, and Mistral."
readme = "README.md"
authors = [
    { name = "Document AI" }
]
requires-python = ">=3.10"
dependencies = [
    # OCR + Layout
    "paddlepaddle>=3.0.0",
    "paddleocr>=2.7.3",

    # PDF / DOCX parsing
    "pdfplumber==0.11.0",
    "PyMuPDF>=1.24.5",
    "python-docx==1.1.2",

    # Chunking & LLM
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-ollama>=0.1.0",
    "tiktoken==0.7.0",

    # Embedding + Vector Store
    "sentence-transformers>=3.0.1",
    "faiss-cpu>=1.9.0",
    "numpy>=2.1.0",
    "scipy>=1.14.0",

    # API layer
    "fastapi==0.111.1",
    "uvicorn[standard]==0.30.3",
    "python-multipart==0.0.9",

    # Utilities
    "tqdm==4.66.4",
    "python-dotenv==1.0.1",
    "loguru==0.7.2",
    "pydantic==2.8.2",

    # Web crawling
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.3",

    # Hybrid Search
    "rank_bm25>=0.2.2",

    # Language detection
    "langdetect>=1.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.27.0",
    "black",
    "flake8",
    "ruff",
]
gpu = [
    "faiss-gpu>=1.9.0",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
